requirements.txt
flask
flask_sqlalchemy
flask_admin
flask_wtf
email-validator
flask_migrate
python-dotenv
Flask-Login
